Re-implement node-daemon run as service and direct run app after merge...
