+++ src/kernel/sys-config/src/app_list.rs
@@ -8,6 +8,10 @@ pub struct SubPkgInfo {
     pub pkg_id: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub docker_image_name:Option<String>,
+    // 留一个口子，以后可以通过这个url下载二进制包
+    // 思考：是不是docker_image_name也可以合并进来？docker://<image_name>:<version>
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub package_url:Option<String>,
     #[serde(flatten)]
     pub configs:HashMap<String,String>,
 
@@ -20,6 +24,8 @@ pub struct AppInfo {
     pub vendor_did: String,
     pub pkg_id: String,
     //service name -> full image url
+    // 命名逻辑:<arch>_<type>_image, 我们把所有的都看作是image，也为了和现在的保持一致
+    // type: direct/docker, 以后可能还有vm
     pub pkg_list: HashMap<String, SubPkgInfo>,
 }
 
@@ -62,6 +68,7 @@ pub struct AppServiceConfig {
     pub user_id: String,
 
     pub docker_image_name : Option<String>,
+    pub direct_image: Option<String>,         // 现在这里只要是Some就可以，以后可以放二进制包的url
     pub data_mount_point: String,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cache_mount_point: Option<String>,
@@ -94,6 +101,7 @@ impl AppServiceConfig {
             app_id: app_config.app_id.clone(),
             user_id:owner_user_id.to_string(),
             docker_image_name: None,//TODO
+            direct_image: None,
             data_mount_point: app_config.data_mount_point.clone(),
             cache_mount_point: app_config.cache_mount_point.clone(),
             local_cache_mount_point: app_config.local_cache_mount_point.clone(),
