+++ src/rootfs/etc/scheduler/nat_ood_and_sn.template.toml
@@ -145,6 +145,10 @@
             },
             "web_pages" :{
                 "pkg_id" : "home-station-web-page"
+            },
+            "amd64_direct_image" :{
+                "pkg_id" : "home-station-web-page",
+                "package_url": "https://web3.buckyos.io/static/home-station-win.zip"
             }
         }
     },
